Explicit classloader for ServiceLoader
By default when loading a service the current thread context class loader will be used. This can be null and in which case the system classloader will be the fallback.
This can be problematic if one uses a JEE container - like Wildfly - where the services will be loaded by the module class loader and tracing won't work.
By using the same classloader that loaded the GhostWriter class, services classes should be available (it fixes the Wildfly issue as well).